requirements.txt
dash
pandas
requests
nba_api
python-dotenv
psycopg2-binary
SQLAlchemy
dash-bootstrap-components>=1.5.0